requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
fonttools==4.50.0
fpdf2==2.7.8
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1
Werkzeug==3.0.1
